+++ package.json
@@ -409,6 +409,7 @@
   },
   "dependencies": {
     "@dcl/schemas": "^5.18.1",
+    "@dcl/sdk": "^7.0.5",
     "@dcl/wearable-preview": "^1.14.0",
     "@types/analytics-node": "^3.1.9",
     "@types/cmd-shim": "^5.0.0",
